+++ libs/tls-disaster-recovery-poc/src/communicator/conn.rs
@@ -342,128 +342,3 @@ impl IncomingConnections {
         self.incoming.contains_key(&peer_id)
     }
 }
-
-// note: duplex connection
-async fn accept_connection(
-    tls_acceptor: tokio_rustls::TlsAcceptor,
-    tcp_stream: TcpStream,
-    allowed_peer_keys: Arc<CommPeers>,
-    message_sender: tokio::sync::mpsc::UnboundedSender<PeerMessage>,
-    cancel: CancellationToken,
-    connections: Arc<Mutex<IncomingConnections>>,
-) -> anyhow::Result<()> {
-    let (stream, conn) = tokio::select! {
-        _ = cancel.cancelled() => {
-            anyhow::bail!("cancelled");
-    }
-        res = async {
-            let mut stream = tls_acceptor.accept(tcp_stream).await?;
-            let common_state = stream.get_ref().1;
-            let peer_pk = mpc_tls::tls::extract_public_key(common_state)?;
-            let Some(peer_id) = allowed_peer_keys.is_allowed(&peer_pk) else {
-                anyhow::bail!("peer is not in list of allowed peers.");
-            };
-            tracing::info!("Performing P2P handshake with: {:?}", peer_id);
-            handshake(&mut stream, constants::HANDSHAKE_TIMEOUT)
-                .await
-                .context("p2p handshake")?;
-            tracing::info!("(incoming) Concluded P2P handshake with: {:?}", peer_id);
-            let cancel_recv = CancellationToken::new();
-            Ok::<_, anyhow::Error>((stream, Connection{peer_id, cancel: cancel_recv}))
-            } => res?,
-    };
-    connections.lock().await.insert(conn.clone());
-    tokio::spawn(recv_loop(stream, conn.cancel, message_sender, conn.peer_id));
-    Ok(())
-}
-
-async fn listen_incoming(
-    tls_acceptor: tokio_rustls::TlsAcceptor,
-    tcp_listener: tokio::net::TcpListener,
-    allowed_peer_keys: Arc<CommPeers>,
-    cancel: CancellationToken,
-    connections: Arc<Mutex<IncomingConnections>>,
-    message_sender: tokio::sync::mpsc::UnboundedSender<PeerMessage>,
-) {
-    loop {
-        let tcp_stream: TcpStream = tokio::select! {
-            _ = cancel.cancelled() => {
-                tracing::info!("listener cancelled; stopping accept loop");
-                break;
-            }
-            res = tcp_listener.accept() => {
-                match res {
-                    Ok((s, addr)) => {
-                        tracing::debug!(%addr, "accepted TCP connection");
-                        s
-                    }
-                    Err(e) => {
-                        tracing::warn!(error=%e, "accept failed; continuing");
-                        continue;
-                    }
-                }
-            }
-        };
-
-        let tls_acceptor = tls_acceptor.clone();
-        let allowed = allowed_peer_keys.clone();
-        let conns = connections.clone();
-        let tx = message_sender.clone();
-        let child = cancel.child_token();
-
-        tokio::spawn(async move {
-            if let Err(e) =
-                accept_connection(tls_acceptor, tcp_stream, allowed, tx, child, conns).await
-            {
-                tracing::warn!(error=%e, "accept_connection failed");
-            }
-        });
-    }
-}
-
-struct Server {
-    server_config: Arc<rustls::server::ServerConfig>,
-    my_port: u16,
-    allowed_peer_keys: Arc<CommPeers>,
-    cancel: CancellationToken,
-    connections: Arc<Mutex<IncomingConnections>>,
-}
-
-impl Server {
-    pub fn new(
-        server_config: Arc<rustls::server::ServerConfig>,
-        my_port: u16,
-        allowed_peer_keys: Arc<CommPeers>,
-        cancel: CancellationToken,
-        connections: Arc<Mutex<IncomingConnections>>,
-    ) -> Self {
-        Server {
-            server_config,
-            my_port,
-            allowed_peer_keys,
-            cancel,
-            connections,
-        }
-    }
-    pub async fn listen(
-        &mut self,
-        message_sender: tokio::sync::mpsc::UnboundedSender<PeerMessage>,
-    ) -> anyhow::Result<()> {
-        let tls_acceptor = tokio_rustls::TlsAcceptor::from(self.server_config.clone());
-        let tcp_listener = tokio::net::TcpListener::bind(SocketAddr::V4(SocketAddrV4::new(
-            Ipv4Addr::new(0, 0, 0, 0),
-            self.my_port,
-        )))
-        .await
-        .context("TCP bind")?;
-        tokio::spawn(listen_incoming(
-            tls_acceptor,
-            tcp_listener,
-            self.allowed_peer_keys.clone(),
-            self.cancel.child_token(),
-            self.connections.clone(),
-            message_sender,
-        ));
-        Ok(())
-    }
-}
