+++ libs/tls-disaster-recovery-poc/src/communicator/connection.rs
@@ -1,0 +1,9 @@
+use tokio_util::sync::CancellationToken;
+
+use crate::types::CommunicatorPeerId;
+
+#[derive(Clone)]
+struct Connection {
+    pub peer_id: CommunicatorPeerId,
+    pub cancel: CancellationToken,
+}
