+++ libs/tls-disaster-recovery-poc/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod conn;
+pub mod communicator;
 pub(crate) mod constants;
 pub(crate) mod handshake;
 pub mod messages;
